+++ assets/javascripts/kitten/components/cards/crowdfunding-card/stories.js
@@ -1,6 +1,12 @@
 import React from 'react'
 import styled from 'styled-components'
-import { withKnobs, text, boolean, number } from '@storybook/addon-knobs'
+import {
+  withKnobs,
+  text,
+  boolean,
+  number,
+  select,
+} from '@storybook/addon-knobs'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
 import { Grid, GridCol } from '../../grid/grid'
@@ -8,12 +14,15 @@ import { CrowdfundingCard as CrowdfundingCardComponent } from './index'
 import { CrowdfundingCardWidget as CrowdfundingCardWidgetComponent } from './widget'
 import { KissKissBankBankIcon } from '../../../components/icons/kisskissbankbank-icon'
 import { Text } from '../../../components/typography/text'
-import { pxToRem } from '../../../helpers/utils/typography'
+import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
+import { STATE_CHOICES } from './stories/project-state'
+import { ScreenConfig } from '../../../constants/screen-config'
+import { parseHtml } from '../../../helpers/utils/parser'
 
 const StoryContainer = ({ children }) => (
   <Container>
     <Grid>
-      <GridCol col-m="6">
+      <GridCol col-m="4">
         <Marger top="5" bottom="5">
           {children}
         </Marger>
@@ -101,3 +110,52 @@ export const CrowdfundingCardWidget = () => (
     />
   </StoryContainer>
 )
+
+const options = Object.keys(STATE_CHOICES)
+
+const StyledText = styled(Text)`
+  display: inline-flex;
+
+  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
+    display: block;
+  }
+`
+
+const InfoContainer = styled.span`
+  font-size: ${stepToRem(-1)};
+  flex-direction: column;
+  display: inline-flex;
+`
+
+const Info = ({ label, text }) => (
+  <InfoContainer>
+    <StyledText tag="strong" weight="bold">
+      {parseHtml(text)}
+    </StyledText>
+    <span>{parseHtml(label)}</span>
+  </InfoContainer>
+)
+
+export const LendopolisCrowdfundingCard = () => {
+  const widgetState = select('widget State', options, options[0])
+
+  return (
+    <StoryContainer>
+      <CrowdfundingCardComponent
+        href={text('Link', '#')}
+        ownerTitle={text('Owner title', 'Urbasolar Energy Ocean Indien 6')}
+        loading={boolean('Loading', false)}
+        stretch={boolean('Stretch', false)}
+        cardTitle={text('Title', 'Centrale solaire Urbasolar Le Port')}
+        cardSubTitle={text('SubTitle', 'Obligation convertible')}
+        info1={text('Info1', <Info label="maturité" text="48 mois" />)}
+        info2={text('Info2', <Info label="sur 157 000 €" text="157 000 €" />)}
+        info3={text('Info3', <Info label="Taux d'intérêt" text="5,0%" />)}
+        progress={number('Progress', 42)}
+        widgetState={STATE_CHOICES[widgetState]}
+        additionalInfo={'Collecte réservée aux départements : 92, 34, 12, 82'}
+        title="Aller sur la collecte …"
+      />
+    </StoryContainer>
+  )
+}