+++ assets/javascripts/kitten/components/cards/crowdfunding-card/styles.js
@@ -107,12 +107,29 @@ export const StyledCrowdfundingCard = styled.div`
   }
 
   .k-CrowdfundingCard__subtitle__subtitleText {
-    line-height: 1;
+    line-height: 1.2;
     flex: 1;
 
     &.k-CrowdfundingCard__subtitle__subtitleText--truncated {
       white-space: nowrap;
       overflow: hidden;
+    }
+  }
+
+  .k-CrowdfundingCard__additionalInfo {
+    line-height: 1;
+    white-space: nowrap;
+    overflow: hidden;
+    display: flex;
+    margin-top: ${pxToRem(10)};
+
+    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
+      padding: 0 ${COMPONENT_GUTTER};
+    }
+
+    > * {
+      overflow: hidden;
+      text-overflow: ellipsis;
     }
   }